Refactor login modal
