+++ src/actions/navbarActions.js
@@ -2,6 +2,6 @@ export const NavbarActionTypes = Object.freeze({
     TOGGLE_LOGIN_MODAL: "TOGGLE_LOGIN_MODAL",
 });
 
-export const toggleLoginModal = () => ({
+export const toggleAuthModal = () => ({
     type: NavbarActionTypes.TOGGLE_LOGIN_MODAL,
 });
