+++ src/reducers/navbarActionsReducer.spec.js
@@ -1,5 +1,5 @@
 import navbarState from "./navbarActionsReducer";
-import { toggleLoginModal } from "../actions/navbarActions";
+import { toggleAuthModal } from "../actions/navbarActions";
 
 describe("Notifications Reducer", () => {
     it("should correctly initialize notifications state", () => {
@@ -7,16 +7,16 @@ describe("Notifications Reducer", () => {
         expect(state).toEqual({ showLoginModal: false });
     });
 
-    it("should toggle login modal when toggleLoginModal action is called", () => {
+    it("should toggle login modal when toggleAuthModal action is called", () => {
         let state = navbarState(
             {
                 showLoginModal: false,
             },
-            toggleLoginModal()
+            toggleAuthModal()
         );
         expect(state.showLoginModal).toBe(true);
 
-        state = navbarState(state, toggleLoginModal());
+        state = navbarState(state, toggleAuthModal());
         expect(state.showLoginModal).toBe(false);
     });
 });
